refactor: move dataset constants to dedicated files
