+++ src/app/data/element-data.ts
@@ -1,0 +1,17 @@
+export interface PeriodicElement {
+  p1: number;
+  position: number;
+  p2: number;
+  mu1: number;
+  mu2: number;
+  g: number;
+  sigma: number;
+  re: number;
+  eo: number;
+  rel: number;
+  relmu: number;
+}
+
+export const ELEMENT_DATA: PeriodicElement[] = [
+  {position: 1, p1: 1000, p2: 100, mu1: 10, mu2: 1, g: 0.98, sigma: 24.5, re: 35, eo: 10, rel: 10, relmu: 10},
+];
