+++ src/app/data/mesh-data.ts
@@ -1,0 +1,19 @@
+export interface MeshTable {
+  lvl: number;
+  mx: number;
+  my: number;
+  mz: number;
+  nx: number;
+  ny: number;
+  nz: number;
+  nel: number;
+  nq2: number;
+  tdof: number;
+}
+
+export const MESH_DATA: MeshTable[] = [
+  {lvl: 1, mx: 16, my: 16, mz: 32, nx: 8, ny: 8, nz: 16, nel: 2048, nq2: 18785, tdof: 83332},
+  {lvl: 2, mx: 32, my: 32, mz: 64, nx: 16, ny: 16, nz: 32, nel: 16384, nq2: 140481, tdof: 627460},
+  {lvl: 3, mx: 64, my: 64, mz: 128, nx: 32, ny: 32, nz: 64, nel: 131072, nq2: 1085825, tdof: 4867588},
+  {lvl: 4, mx: 128, my: 128, mz: 128, nx: 64, ny: 64, nz: 256, nel: 1048576, nq2: 8536833, tdof: 38341636},
+];
